[#7] Extract repo file
